requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
PyMySQL==1.1.0
Werkzeug==2.3.7
cryptography==41.0.7
python-dotenv==1.0.0
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Pillow==10.0.1